Ashok

+++ src/helpers/api.js
@@ -25,6 +25,9 @@ instance.interceptors.response.use(
 		return response;
 	},
 	async error => {
+		if (error.config.url.split("/").at(-1) === "login") {
+			return Promise.reject(error);
+		}
 		if (error.response.status === 401) {
 			localStorage.clear();
 			sessionStorage.clear();
